+++ src/components/Banner.jsx
@@ -3,33 +3,20 @@ import { Link } from 'react-router-dom';
 
 const Banner = (props) => {
     return (
-        props.isDisabled ?
-        null :
-        <Link className={"banner banner--" + props.type} to={"/" + props.id + "/character"}>
-            {
-            props.isCrest ? 
-            <div className="icon--mid relative">
-                <img src={props.avatar[0]} className="icon--mid absolute" />
-                <img src={props.avatar[1]} className="icon--mid absolute" />
-                <img src={props.avatar[2]} className="icon--mid absolute" />
-            </div> 
-            :
-            <img className="rounded" src={props.avatar} />
-            }
-            <div className="banner__text">
-                <h4>{props.isPrefix ? props.title : ""}</h4>
-                <h2 className={props.type == "free-company" ? "free-company__text" : ""}>
-                    {props.name}
-                </h2>
-                {
-                props.type == "free-company" ?
-                <p>{props.title}</p> :
-                <h4>{props.isPrefix ? "" : props.title}</h4>
-                }
-            </div>
-            <h5 className="banner__server">{props.server}</h5>
-        </Link>
+    <Link className={'banner ' + props.type} to={props.link}>
+        {props.avatar}
+        <div className='col justify-center'>
+            <h2>{props.name}</h2>
+            <h3>{props.title}</h3>
+            <p>{props.content}</p>
+        </div>
+        <p style={{marginLeft: 'auto'}}>{props.misc}</p>
+    </Link>
     );
+}
+
+Banner.defaultProps = {
+    link: ""
 }
 
 export default Banner;