+++ src/pages/Settings.jsx
@@ -3,7 +3,8 @@ import { useNavigate } from 'react-router-dom';
 import { useState, useEffect } from 'react';
 import { BsChevronDown, BsChevronUp } from 'react-icons/bs';
 import Button from '../components/utility/Button';
-import CharacterBanner from '../components/CharacterBanner';
+import Banner from '../components/Banner';
+import Divider from '../components/utility/Divider';
 
 const Settings = (props) => {
 
@@ -15,6 +16,20 @@ const Settings = (props) => {
         // Set new content width
         document.documentElement.style.setProperty('--content-width', '30rem');
         props.setShowSearchbar(false);
+
+        // Load reference character if it exists.
+        const referenceCharacter = JSON.parse(localStorage.getItem('reference'));
+        if (referenceCharacter !== null) {
+            setReferenceBanner(() => 
+                <Banner 
+                    type='reference'
+                    avatar={<img src={referenceCharacter.Avatar} className='rounded' />}
+                    name={referenceCharacter.Name}
+                    title={referenceCharacter.Title.Name}
+                    misc={referenceCharacter.Server}
+                />
+            )
+        }
     }, [])
 
     const updateReference = (e) => {
@@ -28,16 +43,20 @@ const Settings = (props) => {
                     .then(data => referenceCharacter = data.Character)
                 
                 console.log(referenceCharacter);
-
-                localStorage.setItem('reference', JSON.stringify(referenceCharacter))
-                props.setReferenceCharacter(referenceCharacter);
-                setReferenceBanner(() => 
-                    <CharacterBanner 
-                        name={referenceCharacter.Name}
-                        title={referenceCharacter.Title.Name}
-                        avatar={referenceCharacter.Avatar}
-                     />
-                )
+                
+                if (referenceCharacter !== null) {
+                    localStorage.setItem('reference', JSON.stringify(referenceCharacter))
+                    props.setReferenceCharacter(referenceCharacter);
+                    setReferenceBanner(() => 
+                        <Banner 
+                            type='reference'
+                            avatar={<img src={referenceCharacter.Avatar} className='rounded' />}
+                            name={referenceCharacter.Name}
+                            title={referenceCharacter.Title.Name}
+                            misc={referenceCharacter.Server}
+                        />
+                    )
+                }   
             }
             requestData(id);
         }
@@ -47,6 +66,7 @@ const Settings = (props) => {
         <div className="settings">
             <div className="settings__form">
                 <h1>Settings</h1>
+                <Divider />
                 <h3>Theme</h3>
                 <p>Sets the visual theme of <b>XIV Tracker</b></p>
                 <div