+++ changelog.md
@@ -9,6 +9,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 
 - Character banner at the beggining of the page.
+- Reference feature. This feature allows the user to save a reference character. Spoiler content will be hidden based on the reference character progress.
+
+### Changed
+
+- Banner component refactor.
 
 
 ## Released